refactor: add new line in end of file. list to set. use BigDecimal, etc...

+++ src/test/java/LottoTests/LottoResultsTests.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -36,7 +37,7 @@ public class LottoResultsTests {
     @DisplayName("로또 수익률 테스트")
     @ParameterizedTest
     @MethodSource("profitTestCases")
-    public void profitTest(LottoResults lottoResults, Double expectedProfit) {
+    public void profitTest(LottoResults lottoResults, BigDecimal expectedProfit) {
         assertThat(lottoResults.profit()).isEqualTo(expectedProfit);
     }
 
@@ -49,11 +50,11 @@ public class LottoResultsTests {
 
     private static Stream<Arguments> profitTestCases() {
         return Stream.of(
-                Arguments.of(lottoResultsList.get(0), 400311.0),
-                Arguments.of(lottoResultsList.get(1), 10.0),
-                Arguments.of(lottoResultsList.get(2), 500.0),
-                Arguments.of(lottoResultsList.get(3), 0.5),
-                Arguments.of(lottoResultsList.get(4), 0.0));
+                Arguments.of(lottoResultsList.get(0), BigDecimal.valueOf(400311.0)),
+                Arguments.of(lottoResultsList.get(1), BigDecimal.valueOf(10.0)),
+                Arguments.of(lottoResultsList.get(2), BigDecimal.valueOf(500.0)),
+                Arguments.of(lottoResultsList.get(3), BigDecimal.valueOf(0.5)),
+                Arguments.of(lottoResultsList.get(4), BigDecimal.valueOf(0.0)));
     }
 
     private static Stream<Arguments> countTestCases() {
